Refactor paragraph classification with a store
